feat: add license
I think this makes it flexible while not having someone just go and like
try to take it all and sell it. I don't care about the code but the
written words take a lot of effort and energy to write!

I'd consider changing it if there's good reason to.