Refresh hello-dragon image and related text
The latest DRGTK version default output has different text and also
has a rotating DragonRuby image. This commit updates image and text.

For the image, I tried to use the same specs as the original image,
but please let me know if it needs to change.

For the text, I added the fact that the image rotates. But, again,
if leaving this out was intentional, please let me know.